moving files to a more logical structure.
