+++ client/routes/workflow/components/workflow-graph/helpers/get-event-connections.js
