+++ client/routes/workflow/components/workflow-graph/helpers/graph-styles.js
